backend/requirements.txt
# Core Django
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
pymongo==4.6.1

# Security & Authentication
bcrypt==4.1.2

# HTTP & Networking
requests==2.31.0

# Production Server (WSGI)
gunicorn==21.2.0

# Environment Variables Management
python-decouple==3.8

# PostgreSQL Support (if migrating from SQLite)
psycopg2-binary==2.9.9

# Monitoring & Logging
whitenoise==6.6.0

# Excel File Processing
pandas==2.1.4
openpyxl==3.1.2
